Еще немного правок
